Fixes visualizer speed bug

+++ Source/Audio/AudioEngine.cs
@@ -203,7 +203,7 @@ namespace WaveTracker.Audio {
                     }
 
                     if (App.VisualizerMode && !rendering)
-                        if (_tickCounter % (SamplesPerTick / (App.Settings.Visualizer.PianoSpeed / (App.Settings.Visualizer.DrawInHighResolution ? 1 : App.Settings.General.ScreenScale))) == 0) {
+                        if (_tickCounter % (int)(SamplesPerTick / ((float)App.Settings.Visualizer.PianoSpeed / (App.Settings.Visualizer.DrawInHighResolution ? 1 : App.Settings.General.ScreenScale))) == 0) {
                             App.Visualizer.RecordChannelStates();
                             //App.visualization.Update();
                         }
